Initial extrusion setup
