+++ gcode_preprocessing/contour_flipping.py
@@ -189,6 +189,9 @@ def flip_layer_on_contours(layer_a, layer_b):
     Returns:
         str: The full text content of layer_b after flipping 
     """
+    # Remove the absolute extrusion values and replace them with relative extrusion
+    layer_a = relative_extrusion(layer_a)
+    layer_b = relative_extrusion(layer_b)
     # Sometimes the first delimiter isn't recognized so we just get rid of the stuff before it and then tack that part
     # back on at the end
     content_a_before, layer_a, content_b_before, layer_b = insert_first_delimiter(layer_a, layer_b)
@@ -245,6 +248,9 @@ def flip_layer_on_contours(layer_a, layer_b):
     
     flipped_text = ''.join(flipped)
     full_text = content_a_before + flipped_text
+    
+    full_text = absolute_extrusion(full_text)
+
     return full_text
 
 def flip_on_contours(text_a, text_b):
